put back most of r59166
git-svn-id: https://svn.r-project.org/R/trunk@59173 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/datetime.c
@@ -268,10 +268,10 @@ static double mktime00 (struct tm *tm)
     /* safety check for unbounded loops */
     if (year0 > 3000) {
 	excess = (int)(year0/2000) - 1;
-	year0 -= (int)(excess * 2000);
+	year0 -= excess * 2000;
     } else if (year0 < 0) {
 	excess = -1 - (int)(-year0/2000);
-	year0 -= (int)(excess * 2000);
+	year0 -= excess * 2000;
     }
 
     for(i = 0; i < tm->tm_mon; i++) day += days_in_month[i];
